🚚 Rename `globalClusterAssociation` to fix typo in file name
