+++ src/db/index.ts
@@ -41,7 +41,7 @@ import flowObject from './models/flowObject';
 import flowObjectType from './models/flowObjectType';
 import form from './models/form';
 import globalCluster from './models/globalCluster';
-import globalClusterAssociation from './models/globalCusterAssociation';
+import globalClusterAssociation from './models/globalClusterAssociation';
 import governingEntity from './models/governingEntity';
 import governingEntityVersion from './models/governingEntityVersion';
 import location from './models/location';
